Mountain Ville - Sprawling Cities
The first steps to the major city, starting with several separate built areas. This is essentially the foundation work for the major city to be built. Built up several separate areas of the map in a similar but unique style. They will all eventually meet when the city grows. Mini-cities are already pretty detailed and large, and look pretty decent so far.